Consolidate support e2e tests for improved performance
Change-type: patch
